Friday 10-09-15 Changes In Class
All links are fixed with combining specs for semi-rigid and pressure
sensitive & for anti-static & expanded.  Marine grade and pvc specs
show up white for semi-rigid.php, expanded.php, and fire-resistant.php.
 scratch resistant was taken off of polyethylene and polypropylene and
it was added to anti-static.php.  It was added to pvc.php at the top
and I am still working on adding it to the footer